Revert "[web] Split test-utils in two files"
This reverts commit 5ef9c478ac4f07e88ec3297cfab1da1617e1b252.
