+++ web/src/App.test.jsx
@@ -21,8 +21,7 @@
 
 import React from "react";
 import { act, screen } from "@testing-library/react";
-import { mockComponent } from "@test-utils/mocks";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender, mockComponent } from "@/test-utils";
 import App from "./App";
 import { createClient } from "@client";
 import { STARTUP, CONFIG, INSTALL } from "@client/phase";
