+++ web/src/components/core/Fieldset.test.jsx
@@ -21,7 +21,7 @@
 
 import React from "react";
 import { screen, within } from "@testing-library/react";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender } from "@/test-utils";
 import { Fieldset } from "@components/core";
 
 const ComplexLegend = () => {
