+++ web/src/components/core/InstallationFinished.test.jsx
@@ -22,7 +22,7 @@
 import React from "react";
 
 import { screen } from "@testing-library/react";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender } from "@/test-utils";
 import { createClient } from "@client";
 
 import InstallationFinished from "./InstallationFinished";
