+++ web/src/components/core/InstallationProgress.test.jsx
@@ -22,8 +22,7 @@
 import React from "react";
 
 import { screen } from "@testing-library/react";
-import { mockComponent } from "@test-utils/mocks";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender, mockComponent } from "@/test-utils";
 
 import InstallationProgress from "./InstallationProgress";
