+++ web/src/components/core/Section.test.jsx
@@ -21,7 +21,7 @@
 
 import React, { useState } from "react";
 import { screen } from "@testing-library/react";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender } from "@/test-utils";
 import { Section } from "@components/core";
 
 describe("Section", () => {
