+++ web/src/components/core/ValidationErrors.test.jsx
@@ -21,7 +21,7 @@
 
 import React from "react";
 import { screen, waitFor } from "@testing-library/react";
-import { plainRender } from "@test-utils/renderers";
+import { plainRender } from "@/test-utils";
 import ValidationErrors from "./ValidationErrors";
 
 describe("when there is a single error", () => {
