+++ web/src/components/layout/LoadingEnvironment.test.jsx
@@ -22,7 +22,7 @@
 import React from "react";
 
 import { screen } from "@testing-library/react";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender } from "@/test-utils";
 
 import LoadingEnvironment from "./LoadingEnvironment";
