+++ web/src/components/overview/StorageSection.test.jsx
@@ -21,8 +21,7 @@
 
 import React from "react";
 import { act, screen, waitFor } from "@testing-library/react";
-import { createCallbackMock, mockComponent } from "@test-utils/mocks";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender, createCallbackMock, mockComponent } from "@/test-utils";
 import { createClient } from "@client";
 import { BUSY, IDLE } from "@client/status";
 import { StorageSection } from "@components/overview";
