+++ web/src/components/storage/DeviceSelector.test.jsx
@@ -21,7 +21,7 @@
 
 import React, { useState } from "react";
 import { screen, waitFor, within } from "@testing-library/react";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender } from "@/test-utils";
 import { DeviceSelector } from "@components/storage";
 
 const availableDevices = [
