+++ web/src/components/storage/ProposalActions.test.jsx
@@ -21,7 +21,7 @@
 
 import React from "react";
 import { screen, waitFor, within, waitForElementToBeRemoved } from "@testing-library/react";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender } from "@/test-utils";
 import { ProposalActions } from "@components/storage";
 
 const actions = [
