+++ web/src/components/storage/ProposalActionsSection.test.jsx
@@ -21,8 +21,7 @@
 
 import React from "react";
 import { screen } from "@testing-library/react";
-import { mockComponent } from "@test-utils/mocks";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender, mockComponent } from "@/test-utils";
 import { ProposalActionsSection } from "@components/storage";
 
 jest.mock("@components/storage/ProposalActions", () => mockComponent("ProposalActions content"));
