+++ web/src/components/storage/ProposalPage.test.jsx
@@ -21,8 +21,7 @@
 
 import React from "react";
 import { screen, waitForElementToBeRemoved } from "@testing-library/react";
-import { mockComponent } from "@test-utils/mocks";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender, mockComponent } from "@/test-utils";
 import { createClient } from "@client";
 import { ProposalPage } from "@components/storage";
