+++ web/src/components/storage/ProposalTargetSection.test.jsx
@@ -21,8 +21,7 @@
 
 import React from "react";
 import { screen, waitFor, within } from "@testing-library/react";
-import { mockComponent } from "@test-utils/mocks";
-import { installerRender } from "@test-utils/renderers";
+import { installerRender, mockComponent } from "@/test-utils";
 import { ProposalTargetSection } from "@components/storage";
 
 const FakeProposalTargetForm = ({ id, onSubmit }) => {
